pyproject.toml
[project]
name = "borea_python"
version = "1.0.5"
authors = [
  { name="David Bauch", email="pypi@borea.dev" },
]
license = { text = "MIT" }
description = "Generate extensible python HTTP client from OpenAPI spec"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "click>=8.0.0,<9.0.0",
    "datamodel-code-generator>=0.28.0,<1.0.0",
    "httpx>=0.20.0,<0.28.0",
    "jinja2>=3.0.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "ruff>=0.2,<0.3",
]
keywords = [
    "OpenAPI",
    "Client",
    "Generator",
    "Borea",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/Borea-dev/python-client-generator"
Issues = "https://github.com/Borea-dev/python-client-generator/issues"

[tool.ruff]
exclude = [
    "*.jinja*",
]

[tool.pdm.scripts]
lint = "ruff check --fix ."
format = "ruff format ."

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/borea_python"]
exclude = [".gitignore"]

# [tool.hatch.metadata.entry-points.console_scripts]
# borea_python = "borea_python.cli:main"
